refactor: Manager 패키지 분리
